Move json to data dir
